[FEAT] : nickname api 뷰 띄우기 / #18
