+++ app/src/main/java/com/example/airbnb/data/ApiFactory.kt
@@ -2,7 +2,6 @@ package com.example.airbnb.data
 
 import android.util.Log
 import com.example.airbnb.BuildConfig.BASE_URL
-import com.example.airbnb.data.api.ExampleApiService
 import com.example.airbnb.data.api.ExploreApiService
 import com.example.airbnb.data.api.WhereApiService
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
@@ -39,7 +38,6 @@ object ApiFactory {
 }
 
 object ServicePool {
-    val exampleService = ApiFactory.create<ExampleApiService>()
     val exploreService = ApiFactory.create<ExploreApiService>()
     val whereService = ApiFactory.create<WhereApiService>()
 }
