+++ app/src/main/java/com/example/airbnb/presentation/where/WhereViewModel.kt
@@ -12,8 +12,12 @@ class WhereViewModel : ViewModel() {
     private val _whereLiveData: MutableLiveData<UiState<List<WhereItem>>> = MutableLiveData()
     val whereLiveData: MutableLiveData<UiState<List<WhereItem>>> = _whereLiveData
 
+    private val _getNickName: MutableLiveData<UiState<String?>> = MutableLiveData()
+    val getNickName: MutableLiveData<UiState<String?>> = _getNickName
+
     init {
         getWhere()
+        getNickName()
     }
 
     private fun getWhere() = viewModelScope.launch {
@@ -34,5 +38,13 @@ class WhereViewModel : ViewModel() {
         }.onFailure {
             Log.d("wv", it.message.toString())
         }
+    }
+
+    private fun getNickName() = viewModelScope.launch {
+        _getNickName.value = UiState.Loading
+        runCatching {
+            ServicePool.whereService.getNickName(1)
+        }.fold({ _getNickName.value = UiState.Success(it.data?.nickname) },
+            { _getNickName.value = UiState.Failure(it.message.toString()) })
     }
 }
